Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -2,4 +2,7 @@
 #include <stdio.h>
 
 
-void main(){int a,b;scanf("%d %d",&a,&b);printf("%s", sqrt(a)==b?"Yes":"No");}
+void main(){
+    double a,b;
+scanf("%f %f",&a,&b);
+printf("%s", sqrt(a)==b?"Yes":"No");}
